Rename addBookCard>showModal, Redesign, Add qs & el for inital edit-svgs
Renamed the function addBookCard to showModal as the purpose of it
was reevaluated. It now just handles when we need to show the modal.

showModal will make modal visible, determine whether it will be an
edit or an add request, and call prepareModal accordingly.

Also added logic to get the initial edit-svg's via querySelectorAll.
Added logic to addEventListeners to them. They will call showModal.

Tested logic to make sure modal title is Add Book vs Edit Book
accordingly.