Adjust one prepareModal call in showModal, change modal css z index
Adjusted the call to prepareModal for the add branch of the showModal
function. The adjustment was to add an empty string as the second
parameter. The reason was because prepareModal expects two parameters
and the add branch was only using one. I wanted the call to be
consistent with the intent of the function.

I also noticed the footer pattern was over the modal so I adjusted the
z index of .modal to be the highest and once again sit on top of the
view when visible.